requirements.txt
requests
python-dotenv
itertools